Power Up Particles
When a player activates a power up, they now have a particle effect
around them! This has the side effect of putting a particle system on
every player ever. If you click on a player, it will show the particle
effect; simply press stop on the playback window that appears in the
bottom right of your scene, and it will stop until you click the player
again. Annoying, but unfortunately necessary (at least until I look
more into it).